Implemented a facade pattern to opt.py
Added an interface that can be used more easily by a geneticist that is not also a software developer. 
New regions to translate can be added, existing region positions can be changed, as can the start position of the vaccine. At any time the user can print out the resulting gene sequence of the vaccine and the virus/vaccine combination